+++ src/main/java/com/masivian/roulette/model/BetBody.java
@@ -1,13 +1,15 @@
 package com.masivian.roulette.model;
 
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class BetBody {
 	private int number;
 	private String color;
-	private long quantity;
+	private BigDecimal quantity;
 	private boolean betNumber;
 	
 }
